+++ src/main/java/com/wdw/wdw/infra/oauth/provider/UserInfoFactoryImpl.java
@@ -1,8 +1,7 @@
-package com.wdw.wdw.config.oauth.provider;
+package com.wdw.wdw.infra.oauth.provider;
 
 import com.wdw.wdw.exception.InvalidProviderType;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 
 import java.util.Map;
